Standardize.
I also gave the learning objectives a quick edit. One thing to watch for: a learning objective should be "bigger" than a single question, so "Define ____" is a sign that you've probably written your LO a bit too fine-grained.

I also added "XX-" chapter prefixes to the names of the R code chunks to make future debugging easier, and set eval=FALSE on one that didn't yield anything useful (the less we evaluate the faster the book builds and the less chance of future errors).